# Zstd encoder limit and raise encoder cache size